Move sidebar into its own component.
